feishu-streamlit-verify/requirements.txt
streamlit>=1.36.0
requests
apscheduler
python-dotenv
